Improve dtype handling in 'SklearnSGDModel'.
- Scikit-Learn 1.3.0 introduced the possibility to use any dtype for
  SGD-based models' weights.
- As a consequence, this commit introduces the optional 'dtype' argument
  to 'SklearnSGDModel.__init__' and '.from_parameters' methods, which is
  only used with `sickit-learn >=1.3` and an un-initialized model.
- Coherently (and for all scikit-learn versions), additional type/dtype
  verification has been implemented under the hood of `_unpack_batch`.

+++ test/functional/test_main.py
@@ -89,7 +89,8 @@ class DeclearnTestCase:
         """Return a Model suitable for the learning task and framework."""
         if self.framework.lower() == "sksgd":
             return SklearnSGDModel.from_parameters(
-                kind=("regressor" if self.kind == "Reg" else "classifier")
+                kind=("regressor" if self.kind == "Reg" else "classifier"),
+                dtype="float32",
             )
         if self.framework.lower() == "tflow":
             return self._build_tflow_model()
